refactor(metadata): remove comicinfo from fields
Seems backwards, but having thought about it, it makes no sense storing
the ComicInfo metadata inside the book struct.

What we should aim for is to use ComicInfo and similar formats to seed
our internal representation of the metadata, rather than for us to look for ways
to store and handle the metadata.

What do we do in case a `.cbz` includes a ComicInfo and
SomeNewSpec.format too? ComicInfo separates the creator role into
writer, penciler, inker, etc etc, as it is a primarily comics focused
format. What if the SomeNewSpec instead comes from a different
perspective in which most of those roles are ran by a single person?
How do we handle mismatch of names? In the database, do we store all of
the files in a json field? Do we decouple them into fields with
standardized names and store the rest?

I don't like those approaches. I would rather we waste some of the
metadata and not store it anywhere, than to have 5000 different tables
with the same information, structured slightly differently.

All of this ignores that ComicInfo as a spec is incredibly unpolished
and attempts of reviving and refurbishing it have been equally as
unsuccessful.